requirements.txt
pandas>=2.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9
camelot-py[cv]>=0.11
tabula-py>=2.9
python-dotenv>=1.0 